feat: Finish opening implementation

+++ internal/opening/repository/repository.go
@@ -6,7 +6,7 @@ import (
 
 type OpeningRepository interface {
 	FindAll() ([]entity.Opening, error)
-	FindById(id uint) (entity.Opening, error)
+	FindById(id uint64) (entity.Opening, error)
 	Create(opening entity.Opening) (entity.Opening, error)
 	Update(opening entity.Opening) (entity.Opening, error)
 	Delete(opening entity.Opening) (entity.Opening, error)
